refactor(web): extract chat UI components from page.tsx
- Extracted `CodeBlock` to `web/src/components/CodeBlock.tsx`
- Extracted `MessageCopyButton` to `web/src/components/MessageCopyButton.tsx`
- Extracted `ChatMessageBubble` to `web/src/components/ChatMessageBubble.tsx`
- Exported `Message` and `MessageThought` types from `chat/page.tsx` for shared use
- Kept UI behavior and styling identical
